Function tmc_langs::check_exercise_updates

pub fn check_exercise_updates(
    client: &TmcClient, 
    projects_dir: &Path
) -> Result<Vec<u32>, LangsError>

Checks the server for any updates for exercises within the given projects directory. Returns the ids of each exercise that can be updated.