Function tmc_langs::prepare_solution

pub fn prepare_solution(
    exercise_path: &Path,
    dest_root: &Path,
) -> Result<(), LangsError>

Note: used by tmc-server. Walks through each given path, processing files and copying them into the destination.

Skips hidden directories, directories that contain a .tmcignore file in their root, as well as files matching patterns defined in FILES_TO_SKIP_ALWAYS and directories and files named private.

Binary files are copied without extra processing, while text files are parsed to remove solution tags and stubs.