Module tmc_langs::progress_reporter

Utility struct for printing progress reports.

Structs

StatusUpdate

The format for all status updates. May contain some data.

Functions

finish_stage

Finishes the current stage.

progress_stage

Progresses the current stage.

start_stage

Starts a new stage.

subscribe

Subscribes to progress reports of type T with callback of type F called every time progress is made with type T.