Struct tmc_langs::TmcClient

pub struct TmcClient(_);

A struct for interacting with the TestMyCode service, including authentication.

Implementations

impl TmcClient

pub fn get_json_from_url<T>(
&self,
url: Url,
params: &[(String, String)]
) -> Result<T, ClientError>where
T: DeserializeOwned,

pub fn download_exercise(
&self,
exercise_id: usize,
target: &Path
) -> Result<(), ClientError>

impl TmcClient

pub fn new(
config_dir: PathBuf,
root_url: String,
client_name: String,
client_version: String
) -> Result<TmcClient, ClientError>

Creates a new TmcClient with the given config directory and root URL.

Errors

This function will return an error if parsing the root_url fails.

Examples

use tmc_client::TmcClient;
use std::path::PathBuf;

let client = TmcClient::new(PathBuf::from("./config"), "https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();

pub fn new_in_config(
root_url: String,
client_name: String,
client_version: String
) -> Result<TmcClient, ClientError>

Creates a new TmcClient with the given root URL. The config directory is set according to dirs::cache_dir.

Errors

This function will return an error if parsing the root_url fails, or if fetching the cache directory fails (see dirs::cache_dir()).

Examples

use tmc_client::TmcClient;

let client = TmcClient::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_ver".to_string()).unwrap();

pub fn set_token(
&mut self,
token: StandardTokenResponse<EmptyExtraTokenFields, BasicTokenType>
) -> Result<(), ClientError>

pub fn authenticate(
&mut self,
client_name: &str,
email: String,
password: String
) -> Result<StandardTokenResponse<EmptyExtraTokenFields, BasicTokenType>, ClientError>

Attempts to log in with the given credentials, returns an error if an authentication token is already present. Username can be the user’s username or email.

Errors

This function will return an error if the client has already been authenticated, if the client_name is malformed and leads to a malformed URL, or if there is some error during the token exchange (see oauth2::Client::excange_password).

Examples

use tmc_client::TmcClient;

let mut client = TmcClient::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();
client.authenticate("client", "user".to_string(), "pass".to_string()).unwrap();

pub fn get_organizations(
&self
) -> Result<Vec<Organization, Global>, ClientError>

Fetches all organizations.

Errors

Returns an error if there’s some problem reaching the API, or if the API returns an error.

pub fn get_organization(
&self,
organization_slug: &str
) -> Result<Organization, ClientError>

Fetches an organization.

Errors

Returns an error if there’s some problem reaching the API, or if the API returns an error.

pub fn get_course_details(
&self,
course_id: usize
) -> Result<CourseDetails, ClientError>

Fetches the course’s information.

Errors

Returns an error if there’s some problem reaching the API, or if the API returns an error.

Examples

use tmc_client::TmcClient;

let client = TmcClient::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();
// authenticate
let course_details = client.get_course_details(600).unwrap();

pub fn get_exercise_details(
&self,
exercise_id: usize
) -> Result<ExerciseDetails, ClientError>

pub fn get_exercises_details(
&self,
exercise_ids: &[usize]
) -> Result<Vec<ExercisesDetails, Global>, ClientError>

pub fn get_course_submissions(
&self,
course_id: usize
) -> Result<Vec<Submission, Global>, ClientError>

pub fn list_courses(
&self,
organization_slug: &str
) -> Result<Vec<Course, Global>, ClientError>

Fetches all courses under the given organization.

Errors

Returns an error if there’s some problem reaching the API, or if the API returns an error.

Examples

use tmc_client::TmcClient;

let client = TmcClient::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();
// authenticate
let courses = client.list_courses("hy").unwrap();

pub fn get_course(&self, course_id: usize) -> Result<CourseData, ClientError>

pub fn get_course_exercises(
&self,
course_id: usize
) -> Result<Vec<CourseExercise, Global>, ClientError>

pub fn paste(
&self,
submission_url: Url,
submission_path: &Path,
paste_message: Option<String>,
locale: Option<Language>
) -> Result<NewSubmission, ClientError>

Sends the given submission as a paste.

Errors

Returns an error if there’s some problem reaching the API, or if the API returns an error.

Examples

use tmc_client::{TmcClient, Language};
use url::Url;
use std::path::Path;

let client = TmcClient::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();
// authenticate
let course_details = client.get_course_details(600).unwrap();
let submission_url = &course_details.exercises[0].return_url;
let submission_url = Url::parse(&submission_url).unwrap();
let new_submission = client.paste(
    submission_url,
    Path::new("./exercises/python/123"),
    Some("my python solution".to_string()),
    Some(Language::Eng)).unwrap();

pub fn send_feedback(
&self,
feedback_url: Url,
feedback: Vec<FeedbackAnswer, Global>
) -> Result<SubmissionFeedbackResponse, ClientError>

Sends feedback.

Errors

Returns an error if there’s some problem reaching the API, or if the API returns an error.

pub fn submit(
&self,
submission_url: Url,
submission_path: &Path,
locale: Option<Language>
) -> Result<NewSubmission, ClientError>

Sends the submission to the server.

Errors

Returns an error if there’s some problem reaching the API, or if the API returns an error. The method compresses the submission and writes it into a temporary archive, which may fail.

pub fn download_old_submission(
&self,
submission_id: usize,
target: &Path
) -> Result<(), ClientError>

pub fn get_exercise_submissions_for_current_user(
&self,
exercise_id: usize
) -> Result<Vec<Submission, Global>, ClientError>

pub fn wait_for_submission(
&self,
submission_url: &str
) -> Result<SubmissionFinished, ClientError>

pub fn get_exercise_updates(
&self,
course_id: usize,
checksums: HashMap<usize, String, RandomState>
) -> Result<UpdateResult, ClientError>

Fetches the course’s exercises from the server, and finds new or updated exercises. If an exercise’s id is not found in the checksum map, it is considered new. If an id is found, it is compared to the current one. If they are different, it is considered updated.

Errors

Returns an error if there’s some problem reaching the API, or if the API returns an error.

Examples

use tmc_client::TmcClient;

let client = TmcClient::new_in_config("https://tmc.mooc.fi".to_string(), "some_client".to_string(), "some_version".to_string()).unwrap();
// authenticate
let mut checksums = std::collections::HashMap::new();
checksums.insert(1234, "exercisechecksum".to_string());
let update_result = client.get_exercise_updates(600, checksums).unwrap();

pub fn mark_review_as_read(
&self,
review_update_url: String
) -> Result<(), ClientError>

Mark the review as read on the server.

Errors

Returns an error if there’s some problem reaching the API, or if the API returns an error.

pub fn get_unread_reviews(
&self,
reviews_url: Url
) -> Result<Vec<Review, Global>, ClientError>

Fetches all reviews.

Errors

Returns an error if there’s some problem reaching the API, or if the API returns an error.

pub fn request_code_review(
&self,
submission_url: Url,
submission_path: &Path,
message_for_reviewer: String,
locale: Option<Language>
) -> Result<NewSubmission, ClientError>

Request code review.

Errors

Returns an error if there’s some problem reaching the API, or if the API returns an error. The method compresses the project and writes a temporary archive, which may fail.

pub fn download_model_solution(
&self,
solution_download_url: Url,
target: &Path
) -> Result<(), ClientError>

Downloads the model solution from the given url.

Errors

Returns an error if there’s some problem reaching the API, or if the API returns an error. The method extracts the downloaded model solution archive, which may fail.

pub fn check_submission(
&self,
submission_url: &str
) -> Result<SubmissionProcessingStatus, ClientError>

Checks the status of a submission on the server.

Errors

Returns an error if the client has not been authenticated, or if there’s some problem reaching the API, or if the API returns an error.