Struct tmc_langs::TmcConfig

pub struct TmcConfig {
    pub location: PathBuf,
    pub projects_dir: PathBuf,
    pub table: Table,
}

The main configuration file. A separate one is used for each client.

Fields
location: PathBufprojects_dir: PathBuftable: Table
Implementations

impl TmcConfig

pub fn load(client_name: &str) -> Result<TmcConfig, LangsError>

Reads or initialises the config for the given client.

pub fn load_from( client_name: &str, path: PathBuf ) -> Result<TmcConfig, LangsError>

Reads or initialises for the client from the given path.

pub fn get_projects_dir(&self) -> &Path

Returns the projects dir.

pub fn set_projects_dir( &mut self, target: PathBuf ) -> Result<PathBuf, LangsError>

Sets the projects dir. Returns the old projects dir.

pub fn get(&self, key: &str) -> Option<&Value>

Fetches a value with the given key.

pub fn insert(&mut self, key: String, value: Value) -> Option<Value>

Inserts a value with the given key and value. Returns the old value, if any.

pub fn remove(&mut self, key: &str) -> Option<Value>

Removes the value with the given key. Returns the removed value, if any.

pub fn save(&mut self) -> Result<(), LangsError>

Saves the config struct to the given path.

pub fn reset(client_name: &str) -> Result<(), LangsError>

Reinitialises the config file.

pub fn get_location(client_name: &str) -> Result<PathBuf, LangsError>