Trait tmc_langs::LanguagePlugin

pub trait LanguagePlugin {
    type StudentFilePolicy: StudentFilePolicy;

    pub const PLUGIN_NAME: &'static str;
    pub const LINE_COMMENT: &'static str;
    pub const BLOCK_COMMENT: Option<(&'static str, &'static str)>;

    pub fn scan_exercise(
        &self, 
        path: &Path, 
        exercise_name: String
    ) -> Result<ExerciseDesc, TmcError>;

    pub fn run_tests_with_timeout(
        &self, 
        path: &Path, 
        timeout: Option<Duration>
    ) -> Result<RunResult, TmcError>;

    pub fn is_exercise_type_correct(path: &Path) -> bool;

    pub fn clean(&self, path: &Path) -> Result<(), TmcError>;

    pub fn get_default_student_file_paths() -> Vec<PathBuf, Global>;

    pub fn get_default_exercise_file_paths() -> Vec<PathBuf, Global>;

    pub fn points_parser(
        i: &str
    ) -> Result<(&str, Vec<&str, Global>), Err<VerboseError<&str>>>;

    pub fn run_tests(&self, path: &Path) -> Result<RunResult, TmcError> { ... }

    pub fn check_code_style(
        &self, 
        _path: &Path, 
        _locale: Language
    ) -> Result<Option<StyleValidationResult>, TmcError> { ... }

    pub fn extract_project(
        compressed_project: impl Read + Seek, 
        target_location: &Path, 
        clean: bool
    ) -> Result<(), TmcError> { ... }

    pub fn extract_student_files(
        compressed_project: impl Read + Seek, 
        target_location: &Path
    ) -> Result<(), TmcError> { ... }

    pub fn find_project_dir_in_zip<R>(
        zip_archive: &mut ZipArchive<R>
    ) -> Result<PathBuf, TmcError>
    where
        R: Read + Seek,
    { ... }

    pub fn get_exercise_packaging_configuration(
        configuration: TmcProjectYml
    ) -> Result<ExercisePackagingConfiguration, TmcError> { ... }

    pub fn get_available_points(
        exercise_path: &Path
    ) -> Result<Vec<String, Global>, TmcError> { ... }
}

The trait that each language plug-in must implement.

These implement the operations needed by the TMC server to support a programming language. These are provided as a library to IDE plug-ins as a convenience. IDE plug-ins often need additional integration work to support a language properly. This interface does NOT attempt to provide everything that an IDE plug-in might need to fully support a language.

Parts of this interface may be called in a TMC sandbox.

Implementations must be thread-safe and preferably fully stateless. Users of this interface are free to cache results if needed.

Associated Types

type StudentFilePolicy: StudentFilePolicy

Loading content...
Associated Constants

pub const PLUGIN_NAME: &'static str

pub const LINE_COMMENT: &'static str

pub const BLOCK_COMMENT: Option<(&'static str, &'static str)>

Loading content...
Required methods

pub fn scan_exercise(
&self,
path: &Path,
exercise_name: String
) -> Result<ExerciseDesc, TmcError>

Produces an exercise description of an exercise directory.

This involves finding the test cases and the points offered by the exercise.

Must return Err if the given path is not a valid exercise directory for this language.

pub fn run_tests_with_timeout(
&self,
path: &Path,
timeout: Option<Duration>
) -> Result<RunResult, TmcError>

Runs the tests for the exercise with the given timeout. Used by run_tests with the timeout from the project config.

pub fn is_exercise_type_correct(path: &Path) -> bool

Tells if there’s a valid exercise in this path.

pub fn clean(&self, path: &Path) -> Result<(), TmcError>

Runs clean command e.g make clean for make or mvn clean for maven.

pub fn get_default_student_file_paths() -> Vec<PathBuf, Global>

pub fn get_default_exercise_file_paths() -> Vec<PathBuf, Global>

pub fn points_parser(
i: &str
) -> Result<(&str, Vec<&str, Global>), Err<VerboseError<&str>>>

A nom parser that recognizes a points annotation and returns the inner points value(s).

For example implementations, see the existing language plugins.

Loading content...
Provided methods

pub fn run_tests(&self, path: &Path) -> Result<RunResult, TmcError>

Runs the tests for the exercise.

pub fn check_code_style(
&self,
_path: &Path,
_locale: Language
) -> Result<Option<StyleValidationResult>, TmcError>

Run checkstyle or similar plugin to project if applicable, no-op by default

pub fn extract_project(
compressed_project: impl Read + Seek,
target_location: &Path,
clean: bool
) -> Result<(), TmcError>

Extract a given archive file containing a compressed project to a target location.

This will overwrite any existing files as long as they are not specified as student files by the language dependent student file policy.

pub fn extract_student_files(
compressed_project: impl Read + Seek,
target_location: &Path
) -> Result<(), TmcError>

Extracts student files from the compressed project. It finds the project dir from the zip and extracts the student files from there. Overwrites all files.

pub fn find_project_dir_in_zip<R>(
zip_archive: &mut ZipArchive<R>
) -> Result<PathBuf, TmcError> where
R: Read + Seek,

Searches the zip for a valid project directory. Note that the returned path may not actually have an entry in the zip. The default implementation tries to find a directory that contains a “src” directory, which may be sufficient for some languages. Ignores everything in a __MACOSX directory.

pub fn get_exercise_packaging_configuration(
configuration: TmcProjectYml
) -> Result<ExercisePackagingConfiguration, TmcError>

Returns configuration which is used to package submission on tmc-server.

pub fn get_available_points(
exercise_path: &Path
) -> Result<Vec<String, Global>, TmcError>

Parses exercise files using Self::LINE_COMMENT and Self::BLOCK_COMMENT to filter out comments and Self::points_parser to parse points from the actual code.

Loading content...