Crate tmc_langs

Re-exports

pub use tmc_client::oauth2;

pub use crate::course_refresher::refresh_course;

Modules

config
Handles the CLI's configuration files and credentials.

course_refresher
Course refresher.

data

file_util
Various utility functions, primarily wrapping the standard library's IO and filesystem functions

warning_reporter
Contains an utility for reporting warnings.

Structs

Course

CourseData

CourseDetails

CourseExercise

ExerciseDesc
A description of an exercise.

ExerciseDetails

ExercisePackagingConfiguration
Represents configuration based on which submission may be packaged.

FeedbackAnswer
Used to respond to feedback questions. See TmcClient::send_feedback.

FileLockGuard

NewSubmission

Organization

Review

RunResult
The result of running an exercise's test suite against a submission.

StyleValidationResult
The result of a style check.

Submission

SubmissionFeedbackResponse

SubmissionFinished

TmcClient
A struct for interacting with the TestMyCode service, including authentication.

UpdateResult

Enums

ClientError

ClientUpdateData
The update data type for the progress reporter.

CommandError
An error caused by a failed attempt to execute an external command.

Language

Traits

LanguagePlugin
The trait that each language plug-in must implement.

Functions

checkstyle

clean

compress_project_to

extract_project

extract_student_files

find_exercise_directories

free_disk_space_megabytes

get_available_points

get_exercise_packaging_configuration

list_local_course_exercises

prepare_solution
Walks through each given path, processing files and copying them into the destination.

prepare_stub

prepare_submission

run_tests

scan_exercise

Type Definitions

Token