Function tmc_langs::download_or_update_course_exercises

pub fn download_or_update_course_exercises(
    client: &TmcClient, 
    projects_dir: &Path, 
    exercises: &[usize], 
    download_template: bool
) -> Result<DownloadResult, LangsError>

Downloads the given exercises, by either downloading the exercise template, updating the exercise or downloading an old submission. If the exercise doesn’t exist on disk yet… if there are previous submissions and download_template is not set, the latest submission is downloaded. otherwise, the exercise template is downloaded. If the exercise exists on disk, it is updated using the course template.