tmc_langs
Trait LanguagePlugin

pub trait LanguagePlugin {
    type StudentFilePolicy: StudentFilePolicy;

    const PLUGIN_NAME: &'static str;
    const DEFAULT_SANDBOX_IMAGE: &'static str;
    const LINE_COMMENT: &'static str;
    const BLOCK_COMMENT: Option<(&'static str, &'static str)>;

    // Required methods
    fn scan_exercise(
        &self,
        path: &Path,
        exercise_name: String,
    ) -> Result<ExerciseDesc, TmcError>;
    fn run_tests_with_timeout(
        &self,
        path: &Path,
        timeout: Option<Duration>,
    ) -> Result<RunResult, TmcError>;
    fn find_project_dir_in_archive<R>(
        archive: &mut Archive<R>,
    ) -> Result<PathBuf, TmcError>
       where R: Read + Seek;
    fn is_exercise_type_correct(path: &Path) -> bool;
    fn clean(&self, path: &Path) -> Result<(), TmcError>;
    fn get_default_student_file_paths() -> Vec<PathBuf>;
    fn get_default_exercise_file_paths() -> Vec<PathBuf>;
    fn points_parser(
        i: &str,
    ) -> Result<(&str, Vec<&str>), Err<VerboseError<&str>>>;

    // Provided methods
    fn run_tests(&self, path: &Path) -> Result<RunResult, TmcError> { ... }
    fn check_code_style(
        &self,
        _path: &Path,
        _locale: Language,
    ) -> Result<Option<StyleValidationResult>, TmcError> { ... }
    fn extract_project<R>(
        archive: &mut Archive<R>,
        target_location: &Path,
        clean: bool,
    ) -> Result<(), TmcError>
       where R: Read + Seek { ... }
    fn extract_student_files(
        compressed_project: impl Read + Seek,
        compression: Compression,
        target_location: &Path,
    ) -> Result<(), TmcError> { ... }
    fn is_archive_type_correct<R>(archive: &mut Archive<R>) -> bool
       where R: Read + Seek { ... }
    fn get_exercise_packaging_configuration(
        path: &Path,
    ) -> Result<ExercisePackagingConfiguration, TmcError> { ... }
    fn get_available_points(
        exercise_path: &Path,
    ) -> Result<Vec<String>, TmcError> { ... }

}

The trait that each language plug-in must implement.

These implement the operations needed by the TMC server to support a programming language. These are provided as a library to IDE plug-ins as a convenience. IDE plug-ins often need additional integration work to support a language properly. This interface does NOT attempt to provide everything that an IDE plug-in might need to fully support a language.

Parts of this interface may be called in a TMC sandbox.

Implementations must be thread-safe and preferably fully stateless. Users of this interface are free to cache results if needed.

Required Associated Constants

const PLUGIN_NAME: &'static str

const DEFAULT_SANDBOX_IMAGE: &'static str

const LINE_COMMENT: &'static str

const BLOCK_COMMENT: Option<(&'static str, &'static str)>

Required Associated Types

type StudentFilePolicy: StudentFilePolicy

Required Methods

fn scan_exercise( &self, path: &Path, exercise_name: String, ) -> Result<ExerciseDesc, TmcError>

Produces an exercise description of an exercise directory.

This involves finding the test cases and the points offered by the exercise.

Must return Err if the given path is not a valid exercise directory for this language.

fn run_tests_with_timeout( &self, path: &Path, timeout: Option<Duration>, ) -> Result<RunResult, TmcError>

Runs the tests for the exercise with the given timeout. Used by run_tests with the timeout from the project config.

fn find_project_dir_in_archive<R>( archive: &mut Archive<R>, ) -> Result<PathBuf, TmcError>
where R: Read + Seek,

Searches the zip for a valid project directory. This function is used to detect the language plugin for the archive, so simply finding “src” is not sufficient, e.g. the Python plugin should check that there is an actual “src/*.py” file inside src. Note that the returned path may not actually have an entry in the zip.

fn is_exercise_type_correct(path: &Path) -> bool

Tells if there’s a valid exercise in this path. Delegates to find_project_dir_in_archive by default. Unlike is_archive_type_correct, only checks the root directory.

fn clean(&self, path: &Path) -> Result<(), TmcError>

Runs clean command e.g make clean for make or mvn clean for maven.

fn get_default_student_file_paths() -> Vec<PathBuf>

fn get_default_exercise_file_paths() -> Vec<PathBuf>

fn points_parser(i: &str) -> Result<(&str, Vec<&str>), Err<VerboseError<&str>>>

A nom parser that recognizes a points annotation and returns the inner points value(s).

For example implementations, see the existing language plugins.

Provided Methods

fn run_tests(&self, path: &Path) -> Result<RunResult, TmcError>

Runs the tests for the exercise.

fn check_code_style( &self, _path: &Path, _locale: Language, ) -> Result<Option<StyleValidationResult>, TmcError>

Run checkstyle or similar plugin to project if applicable, no-op by default

fn extract_project<R>( archive: &mut Archive<R>, target_location: &Path, clean: bool, ) -> Result<(), TmcError>
where R: Read + Seek,

Extract a given archive file containing a compressed project to a target location.

This will overwrite any existing files as long as they are not specified as student files by the language dependent student file policy.

fn extract_student_files( compressed_project: impl Read + Seek, compression: Compression, target_location: &Path, ) -> Result<(), TmcError>

Extracts student files from the compressed project. It finds the project dir from the zip and extracts the student files from there. Overwrites all files.

fn is_archive_type_correct<R>(archive: &mut Archive<R>) -> bool
where R: Read + Seek,

Tells if there’s a valid exercise in this archive. Unlike is_exercise_type_correct, searches the entire archive.

fn get_exercise_packaging_configuration( path: &Path, ) -> Result<ExercisePackagingConfiguration, TmcError>

Returns configuration which is used to package submission on tmc-server.

fn get_available_points(exercise_path: &Path) -> Result<Vec<String>, TmcError>

Parses exercise files using Self::LINE_COMMENT and Self::BLOCK_COMMENT to filter out comments and Self::points_parser to parse points from the actual code.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.